% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagecryptr.R
\name{pagecryptr}
\alias{pagecryptr}
\title{Encrypt and HTML Page}
\usage{
pagecryptr(file, password, out_file = NULL)
}
\arguments{
\item{file}{a character string. path/to/html.file}

\item{password}{a character string. Password that unlocks file contents}

\item{out_file}{optional. file path and name for new file.}
}
\value{
Saves an encrypted HTML file to same path as original file input by appending "-protected.html", unless out_file is specified.
}
\description{
This is copy/pasta from the PageCrypt project by Max Laumeister translated for use as an R function. Makes use of the V8 package to create a javascript environment in R.
}
\examples{
if(interactive()){
 file <- system.file("example.html", package = "pagecryptr")
 pagecryptr(file, "password")
}

}
