% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagecryptr.R
\name{pagecryptr_knit}
\alias{pagecryptr_knit}
\title{Knit Rmd to an encrypted HTML}
\usage{
pagecryptr_knit(input, password)
}
\arguments{
\item{input}{a .Rmd to input to knit to html}

\item{...}{}
}
\value{
an encrypted html file
}
\description{
Knit Rmd to an encrypted HTML
}
