% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagecryptr.R
\name{pagecryptr_yaml_knit}
\alias{pagecryptr_yaml_knit}
\title{For use in the YAML header}
\usage{
pagecryptr_yaml_knit(input, ...)
}
\arguments{
\item{input}{a .Rmd to input to knit to html}

\item{...}{}
}
\value{
an encrypted html file
}
\description{
For use in the YAML header
}
